small improvement in abstra-ref

+++ lib/abstr-ref/@Controller/modify_map_pg.m
@@ -1,7 +1,7 @@
 function modify_map_pg(cont,P_set,u)
     
     set = cont.sets;
-    new_keys = setdiff(P_set,set);
+    new_keys = uint32(setdiff(P_set,set));
 
     for i=1:length(new_keys)
         cont.subcontrollers(new_keys(i)) = u;
